[Aditya Kokashe] : added a program to take height in centimeters and convert it into feet and inches.

+++ gcr-codebase/programming-elements/level-1-practice-programs/HeightConverter.java
@@ -1,8 +1,8 @@
 /* Write a program to take height in centimeters and convert it into feet and inches. */
 
 import java.util.Scanner;
-public class HeightConverter {
-	public static void main(String[] args) {
+public class HeightConverter{
+	public static void main(String[] args){
 
 		Scanner input = new Scanner(System.in);
 		//Taking height in centimeters from user
@@ -11,7 +11,7 @@ public class HeightConverter {
 		//Convert centimeters to inches
 		double totalInches = heightCentimeters / 2.54;
 		//Convert inches to feet
-		double feet = (totalInches / 12);
+		int feet = (int) (totalInches / 12);
 		double inches = totalInches % 12;
 
 		//Display the result
